rename constants => meta module
